Finally Merged the Right way
