+++ CCL Oil Labs Control/Views/EditUsrMenu.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="CCL_Oil_Labs_Control.Menus.EditUsrMenu"
+﻿<Window x:Class="CCL_Oil_Labs_Control.Views.EditUsrMenu"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:CCL_Oil_Labs_Control.Menus"
+        xmlns:local="clr-namespace:CCL_Oil_Labs_Control.Views"
         mc:Ignorable="d"
         Title="" Style="{StaticResource BasicWindow}">
     <Grid Background="WhiteSmoke">
